pyproject.toml
[tool.poetry]
name = "project-tools"
version = "0.1.0"
description = "Here are dependencies for ml project"
authors = ["Anastasia <kalashnikova040800@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
numpy = "^1.22.3"
torch = "^2.1.1"
pandas = "^1.5.0"
matplotlib = "^3.6.0"
torchvision = "^0.16.1"
scikit-learn = "^1.3.2"
pathlib = "^1.0.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.5.0"
tqdm = "^4.66.1"
dvc = "^3.33.3"
dvc-gdrive = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ["py38"]

[tool.isort]
src_paths = ["project-tools"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
extend-ignore = ['E501', 'E203', 'W503', 'B950']
exclude = "__pycache__"
max-line-length = 90
max-complexity = 12
select = "B,C,E,F,W,T4,B9"
